+++ R/CorpusData.R
@@ -429,37 +429,10 @@ CorpusData <- R6::R6Class(
       if (verbose) cli_progress_done()
       
       if (verbose) cli_rule("Check result")
-      if (isTRUE(reload)){
-        
-        cl_delete_corpus(corpus = corpus, registry = registry_dir)
-        if (is.null(cl_find_corpus(corpus = corpus, registry = registry_dir))){
-          if (verbose) cli_alert_success("corpus unloaded")
-        } else {
-          if (verbose) cli_alert_danger("unloading corpus failed")
-        }
-        
-        # Not too intuitive, but this (re)loads the corpus
-        size <- cl_attribute_size(
-          corpus = corpus,
-          attribute = "word",
-          attribute_type = "p",
-          registry = registry_dir
-        )
-        
-        if (typeof(cl_find_corpus(corpus, registry = registry_dir)) == "externalptr"){
-          if (verbose)
-            cli_alert_success("corpus reloaded (size: {.val {size}})")
-        } else {
-          if (verbose) cli_alert_danger("reloading corpus failed (CL representation")
-        }
-        
-        if (!cqp_load_corpus(corpus = corpus, registry = registry_dir)){
-          if (verbose)
-            cli_alert_danger(
-              "reloading corpus failed (CQP representation not available)"
-            )
-        }
-      }
+      
+      if (isTRUE(reload))
+        corpus_reload(corpus = corpus, registry = registry_dir)
+      
       
       p_attrs <- corpus_p_attributes(corpus = corpus, registry = registry_dir)
       if (all(p_attributes %in% p_attrs)){